[#42] Move Image class and Color class to proper package
